Move random to random folder and add readme file
